refactor: replace int32 with int64 to avoid overflow

+++ internal/network/protocol/convert/helpers.go
@@ -9,9 +9,9 @@ import (
 
 func cardToProto(c protocol.CardInfo) *pb.CardInfo {
 	return &pb.CardInfo{
-		Suit:  int32(c.Suit),
-		Rank:  int32(c.Rank),
-		Color: int32(c.Color),
+		Suit:  int64(c.Suit),
+		Rank:  int64(c.Rank),
+		Color: int64(c.Color),
 	}
 }
 
@@ -45,7 +45,7 @@ func playerInfoToProto(p *protocol.PlayerInfo) *pb.PlayerInfo {
 	return &pb.PlayerInfo{
 		Id:         p.ID,
 		Name:       p.Name,
-		Seat:       int32(p.Seat),
+		Seat:       int64(p.Seat),
 		Ready:      p.Ready,
 		IsLandlord: p.IsLandlord,
 		CardsCount: int64(p.CardsCount),
@@ -176,8 +176,8 @@ func roomListItemsToProto(rooms []protocol.RoomListItem) []*pb.RoomListItem {
 	for i, r := range rooms {
 		result[i] = &pb.RoomListItem{
 			RoomCode:    r.RoomCode,
-			PlayerCount: int32(r.PlayerCount),
-			MaxPlayers:  int32(r.MaxPlayers),
+			PlayerCount: int64(r.PlayerCount),
+			MaxPlayers:  int64(r.MaxPlayers),
 		}
 	}
 	return result
